move variants into seperate files
