+++ src/renderer/components/+workloads-pods/details/volumes/variants/scale-io.tsx
@@ -1,0 +1,52 @@
+/**
+ * Copyright (c) OpenLens Authors. All rights reserved.
+ * Licensed under MIT License. See LICENSE in root directory for more information.
+ */
+
+import React from "react";
+import { secretsApi } from "../../../../../../common/k8s-api/endpoints";
+import { DrawerItem } from "../../../../drawer";
+import { LocalRef, VolumeVariantComponent } from "../variant-helpers";
+
+export const ScaleIo: VolumeVariantComponent<"scaleIO"> = (
+  ({ pod, variant: { gateway, system, secretRef, sslEnabled = false, protectionDomain, storagePool, storageMode = "ThinProvisioned", volumeName, fsType = "xfs", readOnly = false }}) => (
+    <>
+      <DrawerItem name="Gateway">
+        {gateway}
+      </DrawerItem>
+      <DrawerItem name="System">
+        {system}
+      </DrawerItem>
+      <LocalRef
+        pod={pod}
+        title="Name"
+        ref={secretRef}
+        api={secretsApi}
+      />
+      <DrawerItem name="SSL Enabled">
+        {sslEnabled.toString()}
+      </DrawerItem>
+      <DrawerItem name="SSL Enabled">
+        {sslEnabled.toString()}
+      </DrawerItem>
+      <DrawerItem name="Protection Domain Name" hidden={!protectionDomain}>
+        {protectionDomain}
+      </DrawerItem>
+      <DrawerItem name="Storage Pool" hidden={!storagePool}>
+        {storagePool}
+      </DrawerItem>
+      <DrawerItem name="Storage Mode" hidden={!storageMode}>
+        {storageMode}
+      </DrawerItem>
+      <DrawerItem name="Volume Name">
+        {volumeName}
+      </DrawerItem>
+      <DrawerItem name="Filesystem Type">
+        {fsType}
+      </DrawerItem>
+      <DrawerItem name="Readonly">
+        {readOnly.toString()}
+      </DrawerItem>
+    </>
+  )
+);
